Move pep8 tests to own file
